Add webPageTabBar prop to TabNavigator and related components
Introduces a new webPageTabBar prop to the tab navigation stack and tab bar components, allowing injection of a web-specific tab bar. Refactors PortalBodyContainer by splitting BottomMenu and WebPageTabBar into separate files and updates their usage in TabNavigator. Removes the now-unnecessary PortalBodyContainer/index.tsx.
