+++ packages/kit/src/provider/Container/PortalBodyContainer/BottomMenu.native.tsx
@@ -1,0 +1,3 @@
+export const BottomMenu = () => {
+  return null;
+};
