--- packages/kit/src/provider/Container/PortalBodyContainer/index.tsx
@@ -1,15 +1,0 @@
-import { memo } from 'react';
-
-import { BottomMenu } from './BottomMenu';
-import { WebPageTabBar } from './WebPageTabBar';
-
-function BasicPortalBodyContainer() {
-  return (
-    <>
-      <WebPageTabBar />
-      <BottomMenu />
-    </>
-  );
-}
-
-export const PortalBodyContainer = memo(BasicPortalBodyContainer);
