+++ packages/kit/src/routes/Tab/Navigator.tsx
@@ -19,6 +19,7 @@ import type { ETabRoutes } from '@onekeyhq/shared/src/routes';
 import { Footer } from '../../components/Footer';
 import { useRouteIsFocused } from '../../hooks/useRouteIsFocused';
 import { BottomMenu } from '../../provider/Container/PortalBodyContainer/BottomMenu';
+import { WebPageTabBar } from '../../provider/Container/PortalBodyContainer/WebPageTabBar';
 import { TabFreezeOnBlurContext } from '../../provider/Container/TabFreezeOnBlurContainer';
 
 import { tabExtraConfig, useTabRouterConfig } from './router';
@@ -88,6 +89,7 @@ export function TabNavigator() {
         extraConfig={isShowWebTabBar ? tabExtraConfig : undefined}
         showTabBar={!(isTabletDetailView && isLandscape)}
         bottomMenu={<BottomMenu />}
+        webPageTabBar={<WebPageTabBar />}
       />
       {platformEnv.isWebDappMode && gtMd ? <Footer /> : null}
       <InPageTabContainer />
